[tf.data] graduate rejection_resample API from experimental to `Dataset`
